About memorystore (#8)
* Create AboutMemorystore.md

Start of instructions how to create a memorystore.js file. This file will preload the app with words. Useful for example for a teacher who wants to create a custom version of LearnWords.

* Update AboutMemorystore.md

Elaborated code to initialize the word objects and load them into the localStorage. Still needs storing of the index, settings and verification of the code.

* Updated AboutMemorystore.md